pipeline: change how Response works
Previously the results data was being help in the Response state with
no parent object. This meant that when the results were updated and had
no data that the results in the Response object would not be cleared
due to how react merged state.

With this change the new empty results will now properly override older
results.

Also contains a small optimisation, moving the props creation out of
the loop.