analytics: fix bug in integration point
Previously id would be set to empty string if no analytics object could
be found on the window. This turned out to be a problem because some
pages and extensions assign values to it, causing the code to mistake
them for ga.

The fix was to specifically set the id to null instead of empty string,
and to check if id had been set before using it.